Library for encoders

+++ Gyro_Encoders/Arduino/Control/Encoder.h
@@ -8,17 +8,17 @@ class Encoder {
 
     public:
             Encoder(int encoderA, int encoderB, long deltaT, int ticksPerRev);
-            int getSpeed(); //returns speed in deg/seg
-            int getDistance(); //return distance rotated in degrees
+            double getSpeed(); //returns speed in deg/seg
+            double getDistance(); //return distance rotated in degrees
             void updateCount();
-            int getTicks(); //returns ticks counts
+            long getTicks(); //returns ticks counts
     private:
             int _encoderA, _encoderB; // encoder pins
             double _degPerTick; //degrees of output shaft rotation per encoder tick
             volatile long _count, _oldCount, _newCount;
             long _deltaT; // in microseconds
             int _lastSpeed;
-            int _totalCount;
+            long _totalCount;
 };
 
 #endif
